requirements/base.txt
beautifulsoup4==4.7.1
celery==5.3.4
dj-database-url==0.5.0
Django==4.2.5
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-ratelimit==2.0.0
geoip2==2.9.0
html2text==2018.1.9
python-decouple==3.1
pytz==2018.9
requests==2.21.0
